refactor(reports): align architecture with Wallet module pattern
