+++ client/src/components/pages/Store/Reports/Sales/PaymentBadge.jsx
