+++ client/src/components/pages/Store/Reports/Sales/index.js
@@ -1,0 +1,4 @@
+export { CloseTurnModal } from "./CloseTurnModal";
+export { DayReport } from "./DayReport";
+export { PaymentBadge } from "./PaymentBadge";
+export { SalesTable } from "./SalesTable";
